requirements.txt
pyserial
influxdb
inky